requirements.txt
datasets==2.11.0
gensim==4.3.2
hdbscan==0.8.33
nltk==3.8.1
numpy==1.24.2
openai==1.51.2
pandas==2.0.0
peft==0.10.0
Pillow==10.4.0
scikit_learn==1.2.2
scipy==1.14.1
torch==2.2.0
torchvision==0.17.0
transformers==4.39.0
umap_learn==0.5.6
